adjust so that estimated time displays in order on submit.
